requirements.txt
# Streamlit UI 프레임워크
streamlit==1.44.0

# 비동기 처리 관련
nest-asyncio==1.6.0
anyio==4.9.0

# LangChain 및 LangGraph 코어
langchain-core==0.3.50
langgraph==0.3.21
langgraph-prebuilt==0.1.7

# LangChain 커뮤니티 및 확장
langchain-community==0.3.20
langchain-mcp-adapters==0.0.5
langchain-teddynote==0.3.45

# MCP 서버 구현
mcp==1.4.1

# 환경 변수 관리
python-dotenv==1.1.0

# Google API 클라이언트 및 인증
google-api-python-client==2.166.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-auth==2.38.0

# Upstage 모델 사용
langchain-upstage==0.6.0
<<<<<<< HEAD
=======

# HTTP 요청 라이브러리
requests
httpx
>>>>>>> mini
